# Optimize wk_coords() for xy