Closes #138; although, this doesn't change any sf-related code. Because the `st_coordinates()` objects are matrices, multiple copies are needed to return the object that wk_coords needs (a data frame). It *is* about 2x faster to do this, but the existing route wasn't all that slow to begin with and the copies require about 3x the memory.

I did keep the optimization for xy that came about as part of this, though, since that's about 3x faster if there are no empties and 2x faster if there are.

``` r
library(wk)

points <- xy(runif(1e6), runif(1e6))
bench::mark(
  wk::wk_coords(points),
  wk:::wk_coords.default(points)
)
#> # A tibble: 2 × 6
#>   expression                          min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                     <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 wk::wk_coords(points)            4.88ms   5.57ms     179.     19.1MB     179.
#> 2 wk:::wk_coords.default(points)  16.75ms  16.96ms      58.9    26.8MB     168.

points_with_empty <- points
points_with_empty[100] <- xy(NA, NA)
bench::mark(
  wk::wk_coords(points_with_empty),
  wk:::wk_coords.default(points_with_empty)
)
#> # A tibble: 2 × 6
#>   expression                                     min  median `itr/sec` mem_alloc
#>   <bch:expr>                                <bch:tm> <bch:t>     <dbl> <bch:byt>
#> 1 wk::wk_coords(points_with_empty)            11.4ms  11.6ms      86.6    61.1MB
#> 2 wk:::wk_coords.default(points_with_empty)   18.2ms  18.4ms      54.0    53.4MB
#> # … with 1 more variable: `gc/sec` <dbl>
```